requirements.txt
python-chess
git+https://github.com/faangbait/normalize_human_name.git@main#egg=formatname
